+++ docs/changelog.md
@@ -5,6 +5,7 @@
 ## v2.9.1 (2024-01-11)
 
 - Inlined one in CoreParserTest.php because it was used only in one place
+- Change IfStatement ast node to a better implementation
 
 ----